Revert "Refactoring of `opaygo_functions` (#138)"
This reverts commit 2a9a94c72c56a5de6b4d8d05f925758101cc8f91.
