+++ firmware/src/helpers.hpp
@@ -38,9 +38,9 @@ void urgeent() {
     credit_formular();
   }
 
-  /* if (Mode_select != 2) {
+  if (Mode_select != 2) {
     STS_keypad();
-  } */
+  }
   warn_now = millis();
   if (c_chek == 1) {
     if (warn_now - warntime > 1500) {
@@ -96,4 +96,49 @@ void urgeent() {
       fault = 0;
     }
   }
+}
+
+uint32_t convertArrayToUint32(int arrayToConvert[9]) {
+  int i = 0;
+  uint32_t uint32 = 0;
+  for (i = 0; i < 9; i++) {
+    uint32 = 10 * uint32 + arrayToConvert[i];
+  }
+  return (uint32);
+}
+
+uint32_t convertByteArrayToUint32(uint8_t arrayToConvert[4]) {
+  uint32_t low1 = arrayToConvert[0];
+  uint32_t low2 = arrayToConvert[1];
+  low2 = low2 << 8;
+  uint32_t high1 = arrayToConvert[2];
+  high1 = high1 << 16;
+  uint32_t high2 = arrayToConvert[3];
+  high2 = high2 << 24;
+  uint32_t result = low1 + low2 + high1 + high2;
+  return (result);
+}
+
+void convertUint32ToUint8Array(uint32_t uint32ToConvert,
+                               uint8_t arrayBytes[4]) {
+  byte low1 = uint32ToConvert;
+  byte low2 = uint32ToConvert >> 8;
+  byte high1 = uint32ToConvert >> 16;
+  byte high2 = uint32ToConvert >> 24;
+  arrayBytes[0] = low1;
+  arrayBytes[1] = low2;
+  arrayBytes[2] = high1;
+  arrayBytes[3] = high2;
+}
+
+void storeUint32InNvram(uint32_t toStore, int address) {
+  uint8_t arrayBytes[4];
+  convertUint32ToUint8Array(toStore, arrayBytes);
+  rtc.writenvram(address, arrayBytes, 4);
+}
+
+uint32_t readUint32FromNvram(int address) {
+  uint8_t readData[4] = {0};
+  rtc.readnvram(readData, 4, address);
+  return (convertByteArrayToUint32(readData));
 }
