+++ firmware/src/main.cpp
@@ -9,13 +9,10 @@
 #include "credit.hpp"
 #include "global_defines.hpp"
 #include "helpers.hpp"
-#include "keypad.hpp"
 #include "lcd_display.hpp"
 #include "lcd_init.hpp"
-#include "led_buzzer.hpp"
 #include "mem_init.hpp"
-#include "mem_operation.hpp"
-#include "opaygo_token.hpp"
+#include "opaygo_functions.hpp"
 #include "power.hpp"
 #include "relay.hpp"
 #include "remote.hpp"
